forms: silence MinGW -Werror warnings (PR #81 CI)
The local mingw-w64 cross-compile (replicating the CI environment)
flagged a handful of warnings that, when -Werror is in scope (some
MSYS2 / mingw-w64 toolchain configs do this implicitly), promote to
errors and fail the build.  Fix them:

  - try_enable_per_monitor_v2 / native_dpi_for / native_dark_mode:
    the FARPROC -> our function-pointer cast triggered both
    -Wpedantic (function-to/from-object-pointer) and
    -Wcast-function-type.  Replace the (T)(void *)raw idiom with
    a `union { FARPROC p; T f; }` pun -- the gcc-recommended way to
    bridge GetProcAddress's return type to a typed pointer with no
    diagnostic.

  - ctl_scrollpanel.c: scrollTo had `if (...) ...; if (...) ...;` on
    a single line which trips -Wmisleading-indentation.  Reformat.

  - ctl_listview.c: the `for (u32 i = 0; i < len && (int)i < (u32)cols; i++)`
    bound mixed signed and unsigned in one expression.  Make `cols`
    a u32 directly so the comparison is clean.

  - ctl_treeview.c: tv_count_subtree was unreferenced (the public
    natives use TVM_GETCOUNT directly) and tripped -Wunused-function.
    Drop it.

  - forms_module.c: native_remove was a Phase 1.x Derma-alias that
    Phase 1.1 unhooked from every meta table.  Dropping it kills
    the matching -Wunused-function.

Verified locally with mingw-w64 13.2 + the CI flags + -Werror (which
the Makefile doesn't enable but is a strict superset of any warning
config the CI might use):

  for f in forms_module.c src/core/*.c src/controls/*.c; do
    x86_64-w64-mingw32-gcc -std=c11 -Wall -Wextra -Wpedantic -Werror \
      -DCANDO_PLATFORM_WINDOWS -D_WIN32_WINNT=0x0600 \
      -DUNICODE -D_UNICODE -c "$f" ...
  done

All TUs compile cleanly.  Linux forms.so + 100 C unit tests still pass.

PR #81.

+++ modules/forms/src/controls/ctl_listview.c
@@ -130,9 +130,9 @@ int native_lv_add_item(CandoVM *vm, int argc, CandoValue *args)
                 /* Array form: each element -> matching column. */
                 CdoObject *arr = cando_bridge_resolve(vm, args[1].as.handle);
                 if (arr) {
-                    u32 len = cdo_array_len(arr);
-                    int cols = (int)SendMessageW(s->hwnd, LVM_GETCOLUMNCOUNT, 0, 0);
-                    for (u32 i = 0; i < len && (int)i < (u32)cols; i++) {
+                    u32 len  = cdo_array_len(arr);
+                    u32 cols = (u32)SendMessageW(s->hwnd, LVM_GETCOLUMNCOUNT, 0, 0);
+                    for (u32 i = 0; i < len && i < cols; i++) {
                         CdoValue cv;
                         if (!cdo_array_rawget_idx(arr, i, &cv)) continue;
                         if (cv.tag == CDO_STRING && cv.as.string) {
